Changed LatticeVal to float (and updated a lot of double references to be LatticeVal references).

+++ src/EWS/physics/Lattice.h
@@ -27,7 +27,7 @@ using std::vector;
 namespace ews {
     namespace physics {
         
-        typedef double LatticeVal;
+        typedef float LatticeVal;
         
         
         /**
